Rename base.py import psyclone.psyir.backend to visitor.py.
Update references to it on other files as well as the name of the
corresponding tests file.
