+++ src/psyclone/psyir/backend/c.py
@@ -41,7 +41,7 @@ it needs to be extended for generating pure C code.
 
 '''
 
-from psyclone.psyir.backend.base import PSyIRVisitor, VisitorError
+from psyclone.psyir.backend.visitor import PSyIRVisitor, VisitorError
 
 
 class CWriter(PSyIRVisitor):
