+++ src/psyclone/tests/psyir/backend/fortran_test.py
@@ -41,7 +41,7 @@ from __future__ import absolute_import
 
 import pytest
 from fparser.common.readfortran import FortranStringReader
-from psyclone.psyir.backend.base import VisitorError
+from psyclone.psyir.backend.visitor import VisitorError
 from psyclone.psyir.backend.fortran import gen_intent, gen_dims, FortranWriter
 from psyclone.psyGen import Symbol, Node, CodeBlock, Container, SymbolTable
 from psyclone.tests.utilities import create_schedule
