+++ src/psyclone/tests/psyir/backend/visitor_test.py
@@ -34,11 +34,11 @@
 # Author R. W. Ford, STFC Daresbury Lab
 # -----------------------------------------------------------------------------
 
-'''Performs pytest tests on the psyclond.psyir.backend.base module'''
+'''Performs pytest tests on the psyclond.psyir.backend.visitor module'''
 
 from __future__ import print_function
 import pytest
-from psyclone.psyir.backend.base import PSyIRVisitor, VisitorError
+from psyclone.psyir.backend.visitor import PSyIRVisitor, VisitorError
 from psyclone.psyGen import Node
